Added exception handling, timeout and window closing event
@Client.java: added registeredStatus and rmiRegistry-Status. So the
client can register again without getting an unhandled exception from
LocateRegistry.createRegistry(1099).
On unregistry the player there is now added the ubinding of the rmi
service.
If player is not getting back, there is now a timeout. But the
automatically win is still missing (on a game with two players).
@ClientGUI.java: added a window closing event - if window closed, the
player gets unregistered (if he is registered).